+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Allow prop `ref` to be passed to `PhoneField`.
+
 ## [0.1.0] - 2020-03-19
 
 ### Added